Add memory_module_middleware
