+++ packages/memory_module/__init__.py
@@ -12,6 +12,7 @@ from memory_module.interfaces.types import (
     UserMessage,
     UserMessageInput,
 )
+from memory_module.utils.teams_bot_middlware import MemoryMiddleware
 
 __all__ = [
     "MemoryModule",
@@ -27,4 +28,5 @@ __all__ = [
     "AssistantMessage",
     "AssistantMessageInput",
     "ShortTermMemoryRetrievalConfig",
+    "MemoryMiddleware",
 ]
